Payment Overview Api Finished

+++ MS3-Back-End/DTOs/ResponseDTOs/Payment/PaymentOverview.cs
@@ -2,9 +2,9 @@
 {
     public class PaymentOverview
     {
-        public double TotalAmount { get; set; }
-        public double FullPayment {  get; set; }
-        public double Installment {  get; set; }
-        public double OverDue {  get; set; }
+        public decimal TotalAmount { get; set; }
+        public decimal FullPayment {  get; set; }
+        public decimal Installment {  get; set; }
+        public decimal OverDue {  get; set; }
     }
 }
